Move breadcrumb spec
