+++ src/@batch-flask/ui/breadcrumbs/breadcrumb-group/index.ts
@@ -1,0 +1,1 @@
+export * from "./breadcrumb-group.component";
