+++ src/@batch-flask/ui/breadcrumbs/breadcrumb.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from "@angular/core";
 import { BrowserModule } from "@angular/platform-browser";
 
-import { BreadcrumbGroupComponent } from "./breadcrumb-group.component";
+import { BreadcrumbComponent } from "./breadcrumb";
+import { BreadcrumbGroupComponent } from "./breadcrumb-group";
 import { BreadcrumbLinkDirective } from "./breadcrumb-link.directive";
-import { BreadcrumbComponent } from "./breadcrumb.component";
 import { BreadcrumbService } from "./breadcrumb.service";
 
 const components = [BreadcrumbGroupComponent, BreadcrumbLinkDirective];
