+++ src/@batch-flask/ui/breadcrumbs/breadcrumb.service.spec.ts
@@ -2,7 +2,7 @@ import { Subject } from "rxjs";
 
 import { Breadcrumb, BreadcrumbService } from "@batch-flask/ui/breadcrumbs";
 import { log } from "@batch-flask/utils";
-import { jobsCrumb, node1Crumb, pool1Crumb, pool1PropertiesCrumb, poolsCrumb } from "./crumbs-data";
+import { jobsCrumb, node1Crumb, pool1Crumb, pool1PropertiesCrumb, poolsCrumb } from "./crumbs-data.spec";
 
 describe("Breadcrumb service", () => {
     let service: BreadcrumbService;
