+++ src/@batch-flask/ui/breadcrumbs/breadcrumb/index.ts
@@ -1,0 +1,1 @@
+export * from "./breadcrumb.component";
