+++ src/@batch-flask/ui/breadcrumbs/index.ts
@@ -1,5 +1,5 @@
-export * from "./breadcrumb-group.component";
+export * from "./breadcrumb-group";
 export * from "./breadcrumb-link.directive";
 export * from "./breadcrumb.service";
-export * from "./breadcrumb-models";
+export * from "./breadcrumbs.model";
 export * from "./breadcrumb.module";
